\name{RcppGraphChi-package}
\alias{RcppGraphChi-package}
\alias{RcppGraphChi}
\docType{package}
\title{
R interface to GraphChi
}
\description{
GraphChi can run very large graph computations on just a single machine,
by using a novel algorithm for processing the graph from disk (SSD or hard drive). 
In some cases GraphChi can solve bigger problems in reasonable time than many other available *distributed* frameworks.
RcppGraphChi provides the R interface to GraphChi through Rcpp and RcppEigen.
}
\details{
\tabular{ll}{
Package: \tab RcppGraphChi\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-11-05\cr
License: \tab GPL-3\cr
}

}
\author{
Qiang Kou, Aapo Kyrola

Maintainer: Qiang Kou <qkou@umail.iu.edu>
}
\references{
GraphChi: http://graphlab.org/projects/graphchi.html
}

\keyword{ package }
