\name{RcppGraphChi-package}
\alias{RcppGraphChi-package}
\alias{RcppGraphChi}
\docType{package}
\title{
R interface to GraphChi
}
\description{
GraphChi can run very large graph computations on just a single machine,
by using a novel algorithm for processing the graph from disk (SSD or hard drive). 
In some cases GraphChi can solve bigger problems in reasonable time than many other available *distributed* frameworks.
RcppGraphChi provides the R interface to GraphChi through Rcpp and RcppEigen.
}
\details{
\tabular{ll}{
Package: \tab RcppGraphChi\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-11-05\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Qiang Kou

Maintainer: Qiang Kou <qkou@umail.iu.edu>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
